Re-implemented the cube solving algorithm so its more reliable.
Also added landmark entities to assist with stitching maps together and cleaned up a few bits here and there.

Parts are also now tagged with the material names that were used for its sides, which can be used in conjunction with a tag editor to help identify textures to be applied to the faces of the generated unions.